requirements-dev.txt
pytest>=7.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9
requests>=2.28.0
beautifulsoup4>=4.9.0
fastapi>=0.95.0
uvicorn>=0.22.0
playwright>=1.40.0